Documents/TODOS/Ingenieria de Sistemas/CALIDAD DE SOFTWARE/CRM - Calidad de Software/nynel-crm/backend/requirements.txt
# Django and core dependencies
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.10
django-environ==0.11.2

# Authentication and Security
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# API Documentation
drf-yasg==1.21.7

# Utilities
Pillow==11.2.1
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1

# Development
django-debug-toolbar==4.2.0
pytest==7.4.4
pytest-django==4.7.0
black==23.12.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0